# Optimize `Hive hash` computation for nested types