--- a/src/main/cpp/src/hive_hash.cu
+++ b/src/main/cpp/src/hive_hash.cu
@@ -580,7 +580,7 @@ std::unique_ptr<cudf::column> hive_hash(cudf::table_view const& input,
   stream.synchronize();
 
   // Compute the hash value for each row
-  thrust::tabulate(rmm::exec_policy(stream),
+  thrust::tabulate(rmm::exec_policy_nosync(stream),
                    output_view.begin<hive_hash_value_t>(),
                    output_view.end<hive_hash_value_t>(),
                    hive_device_row_hasher<hive_hash_function, bool>(nullable,