<!--

Thank you for contributing to RAPIDS Accelerator for Apache Spark!

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please create a draft pull rqeuest
   or prefix the pull request summary with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring additional
   work on top of it then remove any `[WIP]` prefix in the summary and
   restore it from draft status if necessary.

6. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please avoid
   rebasing your branch during the review process, as this causes the context
   of any comments made by reviewers to be lost. If conflicts occur during
   review then they should be resolved by merging into the branch used for
   making the pull request.

Many thanks in advance for your cooperation!

-->
# Main Optimization: 
Flatten nested columns in advance, reducing the size of the `stack_frame`.

# Possible Further Optimization: 
- Make various hash functions share the logic for flattening nested types, but `xxhash64` does not require `_cur_hash`, which presents a challenge for unification.

# Benchmark:
1. Environment: NVIDIA TITAN RTX
2. Sizes: 50MB, 100MB, 500MB, 1GB
- schema: `struct`  with a depth of `max_depth`, with the basic type being `INT32`,` FLOAT32` and `STRING`
```
|  size_bytes  |  max_depth  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |          Diff |   %Diff |  Status  |
|--------------|-------------|------------|-------------|------------|-------------|---------------|---------|----------|
|   50000000   |      1      | 609.823 us |      56.37% | 452.706 us |       8.10% |   -157.117 us | -25.76% |   FAST   |
|  100000000   |      1      | 958.429 us |      32.03% | 606.708 us |      66.13% |   -351.722 us | -36.70% |   FAST   |
|  500000000   |      1      |   3.981 ms |      10.85% |   2.091 ms |       1.65% |  -1889.419 us | -47.46% |   FAST   |
|  1000000000  |      1      |   7.716 ms |       5.89% |   3.929 ms |       0.50% |  -3786.579 us | -49.08% |   FAST   |
|   50000000   |      2      |   1.493 ms |       1.52% | 555.161 us |       5.56% |   -937.916 us | -62.82% |   FAST   |
|  100000000   |      2      |   2.760 ms |       0.61% | 892.648 us |       3.01% |  -1867.525 us | -67.66% |   FAST   |
|  500000000   |      2      |  13.026 ms |       0.39% |   3.539 ms |      16.46% |  -9487.762 us | -72.83% |   FAST   |
|  1000000000  |      2      |  25.811 ms |       0.28% |   6.790 ms |       0.59% | -19021.039 us | -73.69% |   FAST   |
|   50000000   |      4      |   2.530 ms |       7.86% |   1.094 ms |       0.66% |  -1436.823 us | -56.78% |   FAST   |
|  100000000   |      4      |   4.850 ms |       5.51% |   1.946 ms |       1.25% |  -2903.567 us | -59.87% |   FAST   |
|  500000000   |      4      |  23.405 ms |       0.13% |   8.761 ms |       0.47% | -14643.401 us | -62.57% |   FAST   |
|  1000000000  |      4      |  46.649 ms |       0.20% |  17.254 ms |       2.64% | -29395.025 us | -63.01% |   FAST   |
|   50000000   |      8      |   4.468 ms |       0.68% |   2.102 ms |       0.53% |  -2366.352 us | -52.96% |   FAST   |
|  100000000   |      8      |   8.731 ms |       0.44% |   3.953 ms |       0.42% |  -4777.431 us | -54.72% |   FAST   |
|  500000000   |      8      |  42.672 ms |       0.08% |  18.808 ms |       0.39% | -23863.937 us | -55.92% |   FAST   |
|  1000000000  |      8      |  85.191 ms |       0.05% |  37.338 ms |       0.30% | -47853.264 us | -56.17% |   FAST   |
```
- schema: `list` with a depth of `max_depth`, with the basic type being `INT32`
```
|  size_bytes  |  max_depth  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |           Diff |   %Diff |  Status  |
|--------------|-------------|------------|-------------|------------|-------------|----------------|---------|----------|
|   50000000   |      1      | 474.191 us |       5.38% | 467.645 us |      36.75% |      -6.546 us |  -1.38% |   SAME   |
|  100000000   |      1      | 682.398 us |      47.04% | 604.959 us |      65.52% |     -77.439 us | -11.35% |   SAME   |
|  500000000   |      1      |   2.523 ms |      10.46% |   2.630 ms |      11.76% |     106.458 us |   4.22% |   SAME   |
|  1000000000  |      1      |   4.800 ms |       0.20% |   3.370 ms |      10.34% |   -1429.947 us | -29.79% |   FAST   |
|   50000000   |      2      | 890.665 us |       1.14% | 867.720 us |       1.51% |     -22.946 us |  -2.58% |   FAST   |
|  100000000   |      2      |   1.858 ms |       0.55% |   1.832 ms |       1.02% |     -25.131 us |  -1.35% |   FAST   |
|  500000000   |      2      |   9.657 ms |       4.41% |   9.763 ms |       0.44% |     105.982 us |   1.10% |   SLOW   |
|  1000000000  |      2      |  19.158 ms |       2.73% |  19.139 ms |       0.26% |     -19.141 us |  -0.10% |   SAME   |
|   50000000   |      4      |   9.557 ms |       0.41% |   8.215 ms |       0.20% |   -1341.963 us | -14.04% |   FAST   |
|  100000000   |      4      |   9.596 ms |       0.20% |   8.235 ms |       0.21% |   -1360.763 us | -14.18% |   FAST   |
|  500000000   |      4      |  10.212 ms |       0.35% |   8.836 ms |       0.36% |   -1375.577 us | -13.47% |   FAST   |
|  1000000000  |      4      |  18.739 ms |       0.80% |  10.990 ms |       0.25% |   -7748.325 us | -41.35% |   FAST   |
|   50000000   |      8      |   3.643 us |      20.58% |   5.008 us |      12.68% |       1.365 us |  37.48% |   SLOW   |
|  100000000   |      8      |   3.612 us |      19.83% |   4.869 us |      13.84% |       1.257 us |  34.80% |   SLOW   |
|  500000000   |      8      |   19.112 s |        inf% |   18.441 s |        inf% | -671019.531 us |  -3.51% |   SAME   |
|  1000000000  |      8      |   19.252 s |        inf% |   18.581 s |        inf% | -670972.656 us |  -3.49% |   SAME   |
```
- schema: `list` with a depth of `max_depth`, with the basic type being `STRING`
```
|  size_bytes  |  max_depth  |   Ref Time |   Ref Noise |   Cmp Time |   Cmp Noise |         Diff |   %Diff |  Status  |
|--------------|-------------|------------|-------------|------------|-------------|--------------|---------|----------|
|   50000000   |      1      |   1.225 ms |      34.18% |   1.238 ms |      34.57% |    12.566 us |   1.03% |   SAME   |
|  100000000   |      1      |   2.469 ms |       0.32% |   2.474 ms |       0.32% |     5.280 us |   0.21% |   SAME   |
|  500000000   |      1      |  11.670 ms |       0.13% |  11.631 ms |       0.48% |   -39.053 us |  -0.33% |   FAST   |
|  1000000000  |      1      |  23.225 ms |       0.32% |  23.167 ms |       0.24% |   -57.943 us |  -0.25% |   FAST   |
|   50000000   |      2      | 986.634 us |       0.99% | 959.146 us |       1.21% |   -27.489 us |  -2.79% |   FAST   |
|  100000000   |      2      |   4.743 ms |       0.87% |   4.753 ms |       0.84% |     9.306 us |   0.20% |   SAME   |
|  500000000   |      2      |  35.896 ms |       0.59% |  39.882 ms |       0.34% |     3.986 ms |  11.10% |   SLOW   |
|  1000000000  |      2      |  74.205 ms |       0.29% |  82.584 ms |       0.31% |     8.379 ms |  11.29% |   SLOW   |
|   50000000   |      4      |  33.784 ms |       0.11% |  32.820 ms |       0.10% |  -964.448 us |  -2.85% |   FAST   |
|  100000000   |      4      |  34.308 ms |       0.11% |  33.523 ms |       0.22% |  -784.787 us |  -2.29% |   FAST   |
|  500000000   |      4      |  34.769 ms |       0.22% |  34.006 ms |       0.12% |  -763.420 us |  -2.20% |   FAST   |
|  1000000000  |      4      |  35.852 ms |       0.16% |  34.705 ms |       0.11% | -1147.547 us |  -3.20% |   FAST   |
|   50000000   |      8      |   4.155 us |      18.74% |   4.042 us |      18.08% |    -0.113 us |  -2.72% |   SAME   |
|  100000000   |      8      |   3.960 us |      18.35% |   3.965 us |      19.00% |     0.005 us |   0.12% |   SAME   |
|  500000000   |      8      |   4.157 us |      24.39% |   4.064 us |      17.89% |    -0.092 us |  -2.22% |   SAME   |
|  1000000000  |      8      |   52.247 s |        inf% |   52.920 s |        inf% |   673.637 ms |   1.29% |   SAME   |
```