@ttnghia 2024-12-24
```diff
-    [&stream](auto const& col) { return *cudf::column_device_view::create(col, stream); });
```
Remember to never dereference the output from `column_device_view::create` or `table_device_view::create` directly. They return a pointer. If we don't store that pointer, it will be dangled right away thus the reference to them will be invalidiated.