@sperlingxx 2024-12-25
I just came up with some random thought:
How about flattening nested children through a post-order traversal ?  Meanwhile, replaces `first_child_index` with `parent_index`. In addition, adds the virtual root at the end of `flattened_column_views` to store the final result.   
By doing that, we get rid of the stack operations and achieve the hash computation with a simple for-loop. (Because post-order traversal is able to guarantee that all children are handled before their shared parents.)