@res-life 2024-12-25
> I just came up with some random thought: How about flattening nested children through a post-order traversal ? Meanwhile, replaces `first_child_index` with `parent_index`. In addition, adds the virtual root at the end of `flattened_column_views` to store the final result. By doing that, it seems to be able to get rid of the stack operations and achieve the hash computation with a simple for-loop. (Because post-order traversal guarantees that all children are handled before their shared parents.)
> 
> Please correct me if I am wrong. And even if this is doable, I am happy about doing with a follow-up PR since it is a NIT improvement.

Good idea for struct/primitive only types.
But for list type, e.g., list(int) column:
row0: [1,2,3]
row1: [1,2]
row0 and row1 have different number of items, we need a stack for each row to record current child already processed.

For the struct/primitive nested types(without list), we can use this idea.
We can discuss this further more.